OO: Adds explicit cast preventing some compiler warning.
git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@1613 88d5ba13-baa4-4915-ae7a-41304cef7dd7

+++ pkg/kernel/oo/src/ee_semgetvalue.c
@@ -61,7 +61,7 @@ int EE_oo_GetValueSem(SemRefType Sem)
   if (Sem->first == EE_NIL)
     returnvalue = Sem->count;
   else
-    returnvalue = -1;
+    returnvalue = (unsigned int) -1;
 #else
   returnvalue = Sem->count;
 #endif
